requirements.txt
tensorflow-cpu
pyqt5
pillow~=10.2.0
opencv-python~=4.7.0.72
matplotlib~=3.8.3
h5py
numpy~=1.26.4
scipy
requests~=2.31.0
flask
flask_cors
gevent
werkzeug
tqdm~=4.66.2
scikit-learn~=1.4.1.post1
keras~=3.1.1
bokeh~=3.4.0
torch~=2.2.2
torchvision~=0.17.2
pandas~=2.2.1
seaborn~=0.13.2